Implemented partial sync with rollback, batched view download, parallel signature and block verify

+++ daemon/Program.cs
@@ -50,14 +50,20 @@ internal class Program
             Console.WriteLine("Performing full resync");
             var storeDir = Path.Join(dataDir, "store");
 
-            Directory.Delete(storeDir, true);
+            if (Path.Exists(storeDir))
+            {
+                Directory.Delete(storeDir, true);
+            }
         }
 
         if (args.Contains("--force-recreate"))
         {
             var renamedTarget = $"{dataDir}-{DateTimeOffset.Now:yyyyMMddhhmmss}";
-            Directory.Move(dataDir, renamedTarget);
-            Console.WriteLine($"Rename {dataDir} to {renamedTarget}");
+            if (Path.Exists(dataDir))
+            {
+                Directory.Move(dataDir, renamedTarget);
+                Console.WriteLine($"Rename {dataDir} to {renamedTarget}");
+            }
         }
 
         Directory.CreateDirectory(dataDir);
